+++ lib/access_token_agent/invalid_token_type_error.rb
@@ -1,5 +1,7 @@
-class InvalidTokenTypeError < Error
-  def initialize(token_type)
-    super("Expected token_type to be 'bearer', but was '#{token_type}'.")
+module AccessTokenAgent
+  class InvalidTokenTypeError < Error
+    def initialize(token_type)
+      super("Expected token_type to be 'bearer', but was '#{token_type}'.")
+    end
   end
 end